Add padding bottom to the image
Without equal paddings around the image, the bottom two corners looked bad. They did not round like the top two. I managed to mitigate this issue by applying the same amount of padding around the image.

Of course after that I had to remove the top margin of h1 because the space between h1 and the image was bigger.